+++ R/spss_swap.R
@@ -16,6 +16,11 @@ utils::globalVariables(".")
 #' labels from SPSS to proper variable names (e.g.,"gender_sex" or "age_15_25") and sets
 #' them as new column names
 #' @param df tibble read in the \code{\link[haven]{read_sav}} from haven
+#' @param repl_umlaut Default is TRUE. If provided True it replaces umlauts (vowel mutations)
+#' like ä, ö, ü and ß with respective ae, oe, ue, and ss
+#' @param old_itemnames Options are "prepend", "append" or, "remove". Default is "prepend"
+#' With "prepend" the new item names will be prepended by the old item names. E.g., the old item name
+#' was "q0003" and the description "Gender" the new item name will be "q0003_gender"
 #' @return A tibble with more human readable names (old labels as names)
 #'
 #' @author Friedrich-Samuel Taubitz & Bjoern Buedenbender
@@ -26,8 +31,8 @@ utils::globalVariables(".")
 #' @export
 #'
 #' @importFrom labelled var_label
-spss_swap <- function(df) {
-
+spss_swap <- function(df, repl_umlaut = TRUE, old_itemnames = c("remove","prepend","append")) {
+  #TODO Write a mini Vignette for spss_swap
   # Validate user input
   if (!is(df,"tbl_df")) stop("Invalid argument type. df is required to be an spss matrix read with haven::read_sav()")
   # Backup old user input
@@ -39,7 +44,23 @@ spss_swap <- function(df) {
     label <- labelled::var_label(df[,i])
     # Check if  label was supplied
     if(!is.null(unlist(label))){
-      clean_label <- clean_names(label)
+      # Default will be overwritten when old_itemnames is supplied
+      clean_label <- clean_names(label, repl_umlaut = repl_umlaut)
+      # Shall old itemnames be prepend o. appended?
+      if (!missing(old_itemnames) & length(old_itemnames ==1)) {
+        # Prepend old Item Names
+        if(tolower(old_itemnames) == "prepend") {
+          clean_label <- paste(names((df[,i])), # old name
+                               clean_names(label, repl_umlaut = repl_umlaut), # cleaned desc
+                               sep = "_")
+        }
+        if(tolower(old_itemnames) == "append") {
+          clean_label <- paste(clean_names(label, repl_umlaut = repl_umlaut), # cleaned desc
+                               names((df[,i])), # old name
+                               sep = "_")
+        }
+
+      }
     }else{
       clean_label <- names((df[,i]))
     }