+++ NEWS.md
@@ -1,7 +1,13 @@
+# datscience 0.2.3
+
+* Added possibility to maintain old itemnames with spss_swap function
+* Added the possibility of multiple group comparisons to flex_table1 (recommended by Rémi and Taylor)
+
 # datscience 0.2.2
 
 * Removed unnecessary dependencies and replaced them with internal utility functions
 * Added the function flex_table1()
+* Added the function spss_swap to exchange labels / description field of spss with itemnames
 * Wrote an article describing flex_table1()
 
 # datscience 0.2.1